\name{Data_sample}
\alias{Data_sample}
\docType{data}
\title{
Sample data for the PCA based compositive index.
}
\description{
Multivariate data which contains set of positive and negative variables.
}
\usage{data("Data_sample")}

\details{
The multivariate data set comprises ten hypothetical variables ‘A’, ‘B’, ‘C’, ‘D’, ‘E’, ‘F’, ‘G’, ‘H’, ‘I’, and ‘J’ which are carefully chosen indicators of the hypothetical construct for which the composite index is intended to be developed. The variables have different degrees
of importance to define the whole phenomenon represented through the hypothetical
construct. The ten variables need to be assigned relative weights and aggregated to get the
composite index value. Variable ‘E’ has a negative relationship with the construct, meaning
the larger the value of the variable ‘E’, the lesser its effect on the construct. The reverse
consideration is made for the remaining nine positive variables. The variables have different
units of measurement.
}

\examples{
data(Data_sample)
}

